Load campaign data only once
Load data for all campaigns when opening the capaign selection window.
We don't rescan the folders while the window is open so there is no need
to load the lua files multiple times.
Especially the draw code for the map image loads the current campaign
on every draw call to determine what is to be drawn. This also spams the
log with messages such as missing languages.

If we pass the parsed campaign to the mission selection screen we can
also get rid of the invisible column for the filepath.
Finally we can load all ressources (images) in one go.

Fixes #1705